requirements.txt
altair==5.5.0
fastapi==0.115.6
matplotlib==3.9.4
pandas==2.2.3
pdfplumber==0.11.5
pytest==8.3.4
python-pptx==1.0.2
tabula-py==2.10.0
streamlit==1.41.1
uvicorn==0.34.0
requests==2.32.3
